Revert "Create LICENSE"
This reverts commit 63a349d718d3d112cfc8202bcef094d31ab55166.

I'd like to add a license to this since people seem wary to pick
out the bits and pieces they'd like to use. MIT seems like it would
be the most appropriate, but at the same time MIT has the condition
that the copyright notification needs to be included. This is great
for code! Not so much for dotfiles.

I'll either dig through and find a license that says 'do whatever'
or write one out at some point. 