Add a better LICENSE
MIT licensing is great! For code!

However, I don't like the idea of anyone forking or pulling bits and pieces from these dotfiles to have to maintain the licensing of those bits and pieces via MIT. People learn, people use, and people re-use; and I feel the inclusion of any license that isn't 'do whatever you want' would be outside of the spirit of learning or adapting.